IMAP: Don't overwrite full msg in DB with half-loaded msg in memory (#201)
Fixes MIME source not yet downloaded #188

Bug: `downloadCompleted` is true, but the body fields are empty in the DB.

Cause: We do a `save()` of that half-loaded message. And that save is overwriting the complete msg in the DB with the half-loaded one.

It appears there are at least 2 code paths that do that: `IMAPFolder.updateNewFlags()` and `IMAPFolder.messageFlagsChanged()`. They are called e.g. when the read status changes, or flags are added. Both call `await this.saveMsgs(updatedMessages)` (not new msgs). `SQLEMail.save()` then overwrites the existing message. That's the bug.

That's because `SQLEMail.readAll()` reads the headers, but not the body. So, `downloadComplete` will be read as `true`, but the body will still be empty - in memory. Doing `SQLEMail.save()` will the save a broken and inconsistent state back into the DB.

* Add assert protection against this case
* Use `saveMessageWritableProperties()` when saving an incomplete message.
* Be clearer in the function names.