MailZip: Fix empty ZIP files #391 (#396)
The actual problem was a missing `await`; see line 25 of `MailZIP.ts`.

However, this isn't enough to fix the test, as it has other problems:
- It's using the wrong filename to open the ZIP file (so the one it does open is still always empty)
- It's trying to compare a Buffer to a Uint8Array

Investigating the filename issue further, I noticed that `writeZip`'s `filename` is just for locking purposes... but `save` passes the wrong filename, so the locking doesn't actually work. I'm not convinced that locking on the filename is ideal; the other alternative is keying your lock and zip maps on the folder, which would at least avoid an extra call to `getFolderZIPFilePath`, but you might want to use `WeakMap`s if you do that.

While debugging I also confused myself at one point as the test was opening the file saved from the previous test. I've made the test clean up after itself, assuming it succeeds.