Calendar: Edit: Recurrence (#725)
The old code was saving the changes in local variables in the UI, and then pushed it to the event on save. There are at least 2 problems with that:
1. It's against our general idea to a) apply changes directly on the object and b) that changes show up immediately in the rest of the app, whenever possible
2. More importantly, the `let foo =` in RepeatBox.svelte meant that the calendar edit dialog in the sidebar would not refresh and show the values of the previously selected event. This is because `let foo =` is run only when the component is instantiated. Later changes to the properties do *not* re-run that code, and therefore, the variables were not updated to the new data. To fix that, we need to use `$:`, which re-runs the line whenever one of the values changes (as far as Svelte knows, i.e. properties and local vars with assignment, i.e. anywhere on the left side of `=`,  including `foo` in `foo.bar = ...`, and observables with `$`).

However, once I changed to `$:`, the values could not be updated properly by simply writing to them using `bind:value={...}`. Probably, because they were re-calculated from the `event.recurrenceRule` before the latter could be changed.

So, instead of using `bind:`, this is now using `onFooChanged()` user input handlers, which is uglier and more code, and I'm not happy with it, but at least creates a well-defined code path.

This also immediately creates a new `RecurrenceRule` when the user makes changes, and applies it on the event, which means that the change shows in the user's calendar immediately. When he clicks [Revert], we use `unedited` on the master as well to restore the original state.

(Unfortunately, this "immediately see in your calendar" doesn't work for new events yet, simply because they are not yet added to the calendar. This should be fixable.)

This also moves some logic from the UI into the `Event` implementation.